Fix up packaging
