--- src/main/java/website/fgto/Apology.java
@@ -1,11 +1,0 @@
-package website.fgto;
-
-import org.apache.wicket.markup.html.panel.Panel;
-
-public class Apology extends Panel {
-  private static final long serialVersionUID = 1L;
-
-  public Apology(String id) {
-    super(id);
-  }
-}
