+++ src/main/java/website/fgto/royals_of_britain/WicketApplication.java
@@ -1,8 +1,9 @@
-package website.fgto;
+package website.fgto.royals_of_britain;
 
 import org.apache.wicket.markup.html.WebPage;
 import org.apache.wicket.protocol.http.WebApplication;
-import website.fgto.houses.HouseInformationPage;
+import website.fgto.royals_of_britain.pages.houses.HouseInformationPage;
+import website.fgto.royals_of_britain.pages.HomePage;
 
 /**
  * Application object for your web application.
