+++ src/main/java/website/fgto/royals_of_britain/pages/Footer.html
