+++ src/main/java/website/fgto/royals_of_britain/pages/houses/intros/CenelnGabran.java
@@ -1,0 +1,11 @@
+package website.fgto.royals_of_britain.pages.houses.intros;
+
+import org.apache.wicket.markup.html.panel.Panel;
+
+public class CenelnGabran extends Panel {
+  private static final long serialVersionUID = 1L;
+
+  public CenelnGabran(String id) {
+    super(id);
+  }
+}
