+++ src/main/java/website/fgto/royals_of_britain/pages/houses/intros/Stuart.java
@@ -1,4 +1,4 @@
-package website.fgto.houses.intros;
+package website.fgto.royals_of_britain.pages.houses.intros;
 
 import org.apache.wicket.markup.html.panel.Panel;
