+++ src/main/webapp/WEB-INF/web.xml
@@ -21,7 +21,7 @@
 		<filter-class>org.apache.wicket.protocol.http.WicketFilter</filter-class>
 		<init-param>
 			<param-name>applicationClassName</param-name>
-			<param-value>website.fgto.WicketApplication</param-value>
+			<param-value>website.fgto.royals_of_britain.WicketApplication</param-value>
 		</init-param>
 	</filter>
