+++ src/test/java/website/fgto/TestHomePage.java
@@ -3,6 +3,8 @@ package website.fgto;
 import org.apache.wicket.util.tester.WicketTester;
 import org.junit.Before;
 import org.junit.Test;
+import website.fgto.royals_of_britain.pages.HomePage;
+import website.fgto.royals_of_britain.WicketApplication;
 
 /**
  * Simple test using the WicketTester
